+++ docs/sync.js
@@ -34,10 +34,10 @@
  * accept or reject the server's SSL certificate.
  *
  * @property {Realm.Sync~SSLConfiguration} [ssl] - SSL configuration.
- * @property {boolean} [partial] - Whether this Realm should be opened in 'partial synchronization' mode.
- *    Partial synchronisation only synchronizes those objects that match the query specified in contrast
- *    to the normal mode of operation that synchronises all objects in a remote Realm.
- *    **Partial synchronization is a tech preview. Its APIs are subject to change.**
+ * @property {boolean} [full_synchronization] - Whether this Realm should be opened in query-based or full
+ *    synchronization mode. The default is query-based mode which only synchronizes objects that have been subscribed to.
+ *    A fully synchronized Realm will synchronize the entire Realm in the background, irrespectively of the data being
+ *    used or not.
  */
 
 /**